(Nearby Share) Fall back to -build=false data path
In CL:2716970, a workaround was added to use the data files from Nearby
tests from within their fixtures (since fixtures do not yet support data
files). However we did not account for the fact that Tast uses different
data file paths depending on whether or not -build=true or -build=false
is specified in the `tast run` command. This means our workstation runs
are passing (-build=true is default in chroot), while lab runs are
failing (-build=false in the lab). This CL adds a check for the
existence of the -build=true data path, and falls back to the
-build=false data path if it does not exist.

BUG=chromium:1182900
TEST=tast run -build=true nearbyshare.*
TEST=tast run -build=true nearbyshare.* with modification to point to
the prebuilt data path instead to make sure it works in lieu of building
a test image for -build=false.
Exempt-From-Owner-Approval: Simple file existence check that's only for a
temporary data file workaround.

Change-Id: Ie6b2016e8813ceb60ea7bc17c1b66c89bf8a33dc