tast-tests: focus in arc.DragDrop.vm
arc.DragDrop.vm is failing, because since R, UIAutomator's WaitForText
only works for the focused window. It is problematic in arc.DragDrop,
because the test creates a Chrome extension window and it's focused by
default.

With this CL, after performing a drag operation, the test calls
autotestPrivate.setArcAppWindowFocus to focus the ARC window, making
subsequent WaitForTest succeed.

BUG=b:179442415
TEST=tast run rammus-arc-r-DUT arc.DragDrop.vm

Change-Id: Ice5ba8b2694fbdc792d2dc878dc3e1343fd9398e