(Nearby Share) Rename Service package from nearbyshare to nearbyservice.
The package name is the same as the local bundle which means we need to
overload the name during import. This change also makes code easier to
follow when it interacts with the service and the local libraries
together.

nearby_share_service.pb.go is autogenerated.

BUG=chromium:1182879
TEST=nearbyshare.*

Change-Id: I515d5bac4ec6db8972f61e964a0fec26d3b62e1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2727093
Commit-Queue: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>