network.*: move WiFi tests to 'package wifi'
We'd like to group stuff that is mostly specific to the WiFi technology,
rather than general network/connectivity.

On a few (especially those with 'wlan' or 'wifi' prefixes), also rename
to drop redundant info.

I don't rename gRPC services here at the moment, as that's more involved
(I think?) and doesn't reflect directly in the monitoring/naming of
tests.

BUG=b:181060056
TEST=`tast list ${HOST} "wifi.*"`, `tast list ${HOST} "network.*"`
TEST=`tast run ${HOST} wifi.Driver`

Change-Id: I9140354ced95a07027d9906db9646ba3d50c526d