Stop ash.CreateWindows to return connections.
It has been a while since ash.CreateWindows was added, but yet
nobody uses the returned connections, and indeed it is natural
for users when multiple connections are created at once.

Considering that, this CL stops that to return connections.
Instead, the connections are already closed within
ash.CreateWindows.

BUG=none
TEST=fast_build.sh -C

Change-Id: I1d40d9971b71b036d79fd96ca374ed7df0e4098b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2725366
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>