printing: Reduce lpstat log messages
While waiting for IPP-USB configuration, tast repeatedly runs lpstat and
polls for the printer to be added.  Each call except for the final one
will log an identical error due to a non-zero exit status.  The error
message isn't interesting for the test, and the volume causes logs to be
truncated.

Clean this up by moving the lpstat error message into the returned error
message.  The message will then only be logged if the surrounding
testing.Poll loop times out without getting a successful lpstat run.

BUG=b:179413348
TEST=tast run printer.PrintIPPUSB

Change-Id: Ibf4ba9b981839363faaf467213f706d34ad43425