cros/firmware/flashrom_perf.go: Flashrom perf test per region
Implement a E2E test for flashrom to measure the time of execution. Some
tasks such as enrollment are time sensitive and depend on Flashrom.

This gives Flashrom read tests more granularity in their measurements of
the reads of each SPI flash region. This should pave the way to bound
each one to specific time constrains that would otherwise cause test
failure if the deadline has not been meet.

BUG=b:179342220
TEST=`$ tast -verbose run 100.114.231.236 firmware.FlashromPerf.*`
```
 2021/03/05 09:48:40 firmware.FlashromPerf.coreboot  [ PASS ]
 2021/03/05 09:48:40 firmware.FlashromPerf.fmap      [ PASS ]
 2021/03/05 09:48:40 firmware.FlashromPerf.gbb       [ PASS ]
 2021/03/05 09:48:40 firmware.FlashromPerf.rw_elog   [ PASS ]
 2021/03/05 09:48:40 firmware.FlashromPerf.rw_vpd    [ PASS ]
```

Change-Id: I78d449aa3bc5318adacda8d5bfd41a676ccc1a2d