tast-tests: Un-flake the a11y autoclick test.
This change fixes flakes in the a11y.Autoclick test. According to
the tastboard [1], the test commonly fails when attempting to
click the scroll button in the autoclick menu. This step was
failing because we weren't waiting for the scroll button's location
to stabilize before attempting to click it. As a result, we would
move the mouse to the incorrect location, causing the test to fail.

To fix the test, we wait for the scroll button's location to
stabilize before we try to hover and automatically click it.

[1] https://tastboard.corp.google.com/test/a11y.Autoclick

BUG=chromium:1179331
TEST=a11y.Autoclick

Change-Id: I02e2d6d2e82f29f7c74b5e528cdda0106fcb5b84