pkcs11: Fix the lack of scratchpad cleanup in netcertstore
There's no guarantee on the state of the DUT when the test starts, so if the
previous test failed and did not cleanup, the remaining files could affect the
next run.

In order to fix this, cleanup is added at the start of the test so that we can
be sure the scratchpad directory is clean and not cause flakiness.

BUG=b:166639527
TEST=tast run $dut hwsec.Chaps*

Change-Id: Ia989bfb5e422567aae38215ba7d98f525c9ccdd6