(Nearby Share) Add online CB<->Phone tests with large files
Sharing online is an important CUJ, since it allows for an upgrade to
the WebRTC transfer medium which drastically increases sharing speeds. A
30MB file is used, which takes long enough to send that there will be
enough time for the bandwidth upgrade. A .txt file is used instead of an
image, since images are displayed automatically on Android and my local
phone didn't handle opening a 30MB image all that well. In my local
testing a 30MB file took less than 10 seconds to transfer once the
bandwidth upgrade kicked in, but it could occasionally take up to 40
seconds before that happened, so the 1-minute timeout is fairly
generous. The same file over bluetooth would take 2-3 minutes. This CL
only adds online variations for the local tests (CB<->Phone), I will
parameterize the remote CB<->CB tests in a follow-up once remote
fixtures are in.

BUG=chromium:1184028
TEST=tast run <dut> nearbyshare.*txt30mb

Change-Id: I7d67c10880bf2abb1ddaff663beb16fbd852860b