chrome: Rework Chrome options to specify credentials
BACKGROUND

Formerly we had to pass two or more options to perform a GAIA login.
For example, to log in with a supervised account, we had to specify
three options: GAIALogin, Auth, and ParentAuth. This approach is
inconvenient especially when we want to call chrome.New with the same
credentials we picked for a previous chrome.New call with
chrome.AuthPool option.

SUMMARY OF CHANGES

This patch exposes the Creds struct that holds complete credentials,
and reworks chrome.Chrome APIs to work with Creds.

Auth, ParentAuth, Contacts options are now deprecated. Instead,
GAIALogin option now accepts Creds, which allows users to specify both
login mode and credentials at once. For rare use cases where users
want to use non-default credentials for fake logins, a new option
FakeLogin is added.

  chrome.New(ctx, chrome.GAIALogin(chrome.Creds{ ... }))
  chrome.New(ctx, chrome.FakeLogin(chrome.Creds{ ... }))

AuthPool is removed in favor of GAIALoginPool. There are only a few
callers, so we just rewrite all of them to use the new option.

  chrome.New(ctx, chrome.GAIALoginPool(s.RequiredVar(...)))

Users can call Chrome.Creds to obtain complete credentials used for
login. This allows re-logging into the same user profile easily when
a credentials pool is used for the first login. Example:

  cr, err := chrome.New(ctx, chrome.GAIALoginPool(...))
  if err != nil { ... }
  creds := cr.Creds() // save creds
  cr.Close()

  cr, err := chrome.New(ctx, chrome.GAIALogin(creds))
  ...

BUG=chromium:1184434
TEST=tast run $DUT ui.ChromeLoginGAIA

Change-Id: I972f6f454a1e62d5982c1a5e2570eb2b6f19ed59