Make VideoCUJ mute behind a "mute" var
Mute is mainly to not annoy developers and not essential for the
test. Put it behind a "mute" runtime var so that developer could
keep the test quiet. But leave the audio on when running on bots
so that "crastestclient.Mute()" failure does not prevent the
test from running..

BUG=chromium:1183926
TEST=tast run $NOCTURNE ui.VideoCUJ.clamshell

Change-Id: I5f67afeb4af55582f68c8287990007c496a0d826
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2740476
Tested-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>