tast-tests: Add PlayDRM video tests for HW DRM protected video
These new tests are for playing back HW DRM protected video and
verifying all the supported codec + encryption scheme combinations.

They use MSE/EME w/ Shaka Player to playback the content. The content
has the first 5 seconds as clear and the rest of the content is
encrypted w/ the technique matching the test case. The clips are 16
seconds long.

After playback is done, it goes fullscreen with the video and then takes
a screenshot to verify the vast majority of the screen is solid black.
With HW DRM, video screenshots will be solid black. It then also uses
DevTools to verify the video pipeline is set up in a way that maps to HW
DRM (i.e. video is encrypted, HW decoder is used and a decrypting
demuxer is not used).

BUG=b:155501161
TEST=tast run DUT video.PlayDRM.* on volteer

Change-Id: Iccce44d2e300ae3a626280caa467faf9530ed8ab