Tast: Make wifi.CheckIntelSARTable critical again
wifi.CheckIntelSARTable was dropped to informational because a
bad SAR table on Meep was causing failures. This CL makes the test skip
on Meep and re-enables the test as critical.

It does not appear that there are any other devices causing similar
errors, so blocking on meep should be sufficient.

BUG=b:180133528
TEST=tast run ${DUT} wifi.CheckIntelSARTable.informational
runs on Meep
tast run ${DUT} wifi.CheckIntelSARTable
skips on Meep

Change-Id: Ic2c39a6bc38caa4d2ac3f00f72fb017d75fbd0f0