Tast: Refactor common Family Link login code into fixture
Now that several tests log in as a Family Link user in Tast, we should
refactor common code into a fixture so that the tests can share. This
CL is a re-land of crrev/c/2728562. I added chrome.Lock() and
chrome.Unlock() to fix failures when multiple tests run concurrently.

BUG=chromium:1176991
TEST=Ran all affected Tast tests

Change-Id: Ia53d25ea926a6cb315ddef188a26b622b3653b10