wifi: Parameterize ChannelScanDwellTime
As discussed on b/181294019 we'd like to test dwell time
on several distinct WiFi channels. This CL parameterizes
the test to run on several channels with distinct min/max
dwell time criteria for each, but doesn't (yet) tighten
the range of acceptable dwell times:

    Channel 1:  Typically 100ms, Will Accepts 5-250ms
    Channel 9:  Typically 55ms,  Will Accepts 5-250ms
    Channel 36: Typically 10ms,  Will Accepts 5-250ms

The performance metric is also renamed to "dwell_time_ch1"
(or 9, or 36) instead of "dwell_time_single_channel_scan".

BUG=b:181294019
TEST=tast run $DUT 'wifi.ChannelScanDwellTime.*' on Kohaku

Change-Id: I55a2918703c2569f6d32c664b9685dc451a36cc2