tast-tests: Drop GAIA ID parameter on fake logins
It is rarely necessary to specify an explicit GAIA ID when logging in
with a fake account. In fact, most of tests specify hard-coded GAIA IDs
such as "gaia-id" and don't use ID for post-login verifications.

This patch removes explicit GAIA ID from chrome.FakeLogin arguments.
When it is missing, use the default GAIA ID.

BUG=chromium:1186043
TEST=fast_build.sh -T
TEST=tast run $DUT ui.ChromeLogin

Change-Id: I20ab4688cc68523187a056d760d9a42f20f0efab