arc: SignerValidation
This test verifies that ARC signer will pass for this build by passing
signing flow with test dev keys. This is intented to be used as critical
test for Android PFQ to prevent scenarious when Android change may easy
break signer step that leads to long-lasting problem with ARC and
ChromeOS testing.

BUG=b:170344735
TEST=Locally, simulate various problems with ARC signing and test fails.
     CQ presubmit with informational attribute not set to force it
     passing on various bots.

Change-Id: I614d69c2ad8515449a0f8426c18fff63c93d5abd