Eventlog: Add ServiceDep for BiosService
BiosService is used by firmware.Pre. As a result, the firmware.Eventlog
test was failing at precondition-time, because it did not specify a
ServiceDep on BiosService. See results on Tastboard, with a 100% failure
rate: http://tastboard/test/firmware.Eventlog

This CL adds BiosService as a ServiceDep on the firmware.Eventlog test.

BUG=b:174800291
TEST=firmware.Eventlog

Change-Id: I871c51c3a876b2de1323229b7f0561df4809eeff