firmware.Reporters: Move r.Now to new utils.go
r.Now was defined in command_output.go. But that wasn't an appropriate
home for it, since command_output.go was more intended for running
arbitrary DUT commands and collecting output.

This CL adds a new file, utils.go, for miscellaneous utility reporters;
and it moves r.Now() to that file.

BUG=b:174800291
TEST=firmware.Eventlog

Change-Id: I11b8f642f1b71339265d4087676457ecb9ac7a18