Tast: skip CheckIntelSARTable on Meep and Mimrock
Meep and Mimrock have bad GEO SAR tables, which is causing
wifi.CheckIntelSARTable to fail. Add them to 'badModels' so they only
run on the informational version of the test.

BUG=b:180133528
TEST=tast run ${DUT} network.WifiCheckIntelSARTable
tast run ${DUT} network.WifiCheckIntelSARTable.informational
The informational version runs on meep and mimrock and the
non-informational version skips on meep and mimrock

Change-Id: Ie838befdb03b4479dc7e9208a2daa69ee8c0b034