session: Migrate from User to NormalizedUser
In most cases this is correct. In this test we compare the login
username to the username in a protobuf retrieved over DBus and it's
unclear if the username there is normalized or not. Just to be safe,
put this in a seperate CL from the changes we are confident in.

BUG=chromium:358427
TEST=CQ

Disallow-Recycled-Builds: test-failures
Change-Id: I464a587a19c5d3b95e9e4404682c77dcf17e87dc