session: Migrate from User to NormalizedUser
In most cases this is correct. In this test we compare the login
username to the username in a protobuf retrieved over DBus and it's
unclear if the username there is normalized or not. Just to be safe,
put this in a separate CL from the changes we are confident in.

BUG=chromium:358427
TEST=CQ

Disallow-Recycled-Builds: test-failures
Change-Id: I464a587a19c5d3b95e9e4404682c77dcf17e87dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2748212
Tested-by: Fergus Dall <sidereal@google.com>
Auto-Submit: Fergus Dall <sidereal@google.com>
Reviewed-by: Olya Kalitova <okalitova@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>