Adding a delay between toggling to fullscreen
When toggling to fullscreen and back to normal, sometimes because of
quick toggling, the activity stuck in a state between fullscreen and
normal. The activity boundaries are changed to normal but the black
background covers the entire screen.
Polling and Waiting don't help when the activity goes into this
state. The only solution is to avoid the activity to go into this
state. By adding one second delay this can be easily fixed

BUG=b:170160025
TEST=tast run $DUT arc.WMResizableClamshell

Change-Id: I5f17b1094fc5b50a241084f9693c995bcddc8a4c