chrome: Enforce immutability of Config
Config is meant to be immutable after construction in NewConfig.
However this property is not enforced and may be broken by future
changes.

This patch modifies Config to provide access to its values only by
getters, ensuring immutability.

This is a refactoring change to harden the design. No functional
change is expected.

BUG=None
TEST=fast_build.sh -T
TEST=tast run $DUT ui.ChromeLogin ui.ChromeLoginGAIA

Change-Id: Ib82e52f2422b82831613b2af2de3478778905699