graphics: refactor GLBench to its own package
The current design use if/else to determine if glbench is running in
Cros/Crostini.
This approach is hard to maintain if we want to support more environment.

This CL refactors the logic into its own package and introduces an
interface for glbench for future environment.

BUG=b:175636117
TEST=tast run $DUT graphics.GLBench.*hasty*

Change-Id: I10edc0cdae0b455035c047ea83fa7d719e17efb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2886373
Auto-Submit: Po-Hsien Wang <pwang@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>