(Nearby Share) Add a delay when re-adding account to Android
Add a small delay between removing the Android device's GAIA account and
readding it. Re-adding it immediately after is flaky and causes fixture
failures, but waiting an additional second seems to fix it. This will
clean up the current test results, and should hold us over until we find
a better way to do this.

BUG=chromium:1185918
TEST=tast run <dut> nearbyshare.*Phone*
Exempt-From-Owner-Approval: just adding a sleep as a temporary fix

Change-Id: Ib743cec3620eeae6497714194a80da841123516e