policy: Add UserAvatarImage test
Add policy.UserAvatarImage tast test. The test checks the behavior when
the policy is set and when is not set. It implements the steps from the
manual test[1].

Note: even when the policy is set, the user can still click on an
avatar image and the user image preview will be changed. However, the
setting is not applied after closing the window (the avatar will still
be the value from the policy).

Steps:
1- Open chrome://os-settings/changePicture page and click on Cat image.
2- Take a screenshot of the user image preview (should be Cat avatar).
3- Close the window and open the changePicture page again.
4- Take a screenshot of the user image preview.
5- If the policy is set, the the second screenshot in step #4 should
match the policy provided image (red image), so compare the red pixels
percentage and check if it is greater than or equal to 95%.
6- If the policy is not set, the second screenshot in step #4 should
match the first one in step #2, so count the similarity percentage
between the two images and check if it is greater than or equal to 95%.

[1] https://testtracker.googleplex.com/testplans/testcase/detail/799539

BUG=chromium:1185195
TEST=tast run <ip> policy.UserAvatarImage

Change-Id: Ib8833af6f3faee47a441f9f82138bca95fc8747c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2755300
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Mahmoud Gawad <mgawad@google.com>
Tested-by: Mahmoud Gawad <mgawad@google.com>