tast-tests: adjust tablet hotseat swipe position.
Hotseat position returned from chrome.autotestPrivate.getShelfUIInfoForState
might be out of the touch screen boundary, which will cause problem when
use these positions to do the swipe to show hotseat. This CL adjust the
hotseat start and end position before swipe to make sure they are in the
touch boundary.

BUG=chromium:1188301
TEST=./fast_build.sh -T

Change-Id: I32eea7763839ff202f63d9ae82920a58b8743f8f