tast-tests: Fix failure of PowerVideoDecodePerf on DUT without battery
This CL adds _nobatterymetrics variants of PowerVideoDecodePerf tests.
This CL also utilizes the Tast parameter generator to generate params
of the test. The parameterization logic is in
power_video_encode_perf_test.go and it automatically generates params
of the test on its unittest.

BUG=chromium:1065846,b:171533992
TEST=~/trunk/src/platform/tast/tools/go.sh test \
     chromiumos/tast/local/bundles/cros/arc
TEST=tast run $NO_BATTERY_DUT arc.PowerVideoDecodePerf* PASSed/SKIPped

Change-Id: Iaec51eca687297e232e4b2be1075a3d4b0e30fae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2759609
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>