wifi.Reset: Sync the ping config with Autotest.
We don't yet have a good way to wait for reset done. Sync the ping
interval = 1s with Autotest so that we can gain a little more robustness
from the longer interval. Also add log for ping results so that we can
know better how reset can affect the ping when needed.

BUG=b:175602523
TEST=`tast run wifi.Reset` on volteer

Change-Id: Ibd69a2e08c1ad67f6f8d641dc578fdf787c17c07