(Nearby Share) Add 'Some contacts' fixtures
Add fixtures for 'Some contacts' visibility tests. In these fixtures,
the connected Android device will be set as an allowed contact, enabling
the Android device to find the CrOS device as an in-contact receiver.
Also added a fixture for tests that check the negative behavior (CrOS
device not discoverable when Android contact is not selected).

To accommodate local test running on personal OTAs, the runtime vars have
been slightly modified to make 'Some contacts' tests work. The vars for
running with personal OTAs should now look like:
  -var=skipAndroidLogin=true (don't delete and readd accounts)
  -var=android_username="<android_user>" (provide the username so we can
  add it as a selected contact)
  -var=cros_username="<cros user>" (login to CrOS device with this user)
  -var=cros_password="<cros pass>"

BUG=chromium:1185940
TEST=tast run <dut> nearbyshare.*
TEST=tast run <dut> nearbyshare.*somecontacts* (next CL in chain)

Change-Id: Id3bf49bed1d3deeecd5a8b18bb57f4cbbd1e68bf