(Nearby Share) Share common code to common package.
Tast recently forbids importing local libraries in remote tests. We have a
bunch of places that did that. This CL refactors the common stuff into
the common package and updates everywhere that was using them.

nearbysetup and nearbytestutils have moved to common.
Most of the constants at the local nearbyshare have moved too. There was
one that was only used in cros_receive_ui so I just moved it to that
file. Instead of moving filesapp.DownloadsPath and updating a million
callers I just added a const to testutils.

BUG=chromium:1186524
TEST=nearbyshare.* and repo upload hooks.

Change-Id: I1406327d6424b2efd42535c61144e01f3bf8aa7d