network: VPNConnect: Extend waiting timeout for Ethernet online
Currently we are waiting for 15 seconds before the physical Ethernet
service becomes online. It seems not a large enough value for some
cases, and most of the flakiness of the test is caused by that.

This patch extend this value to the same value used in EthernetStaticIP
test.

Also see crrev.com/c/2347395 and the bug link for more details.

BUG=b:165460724
TEST=None

Change-Id: I88ba4d53e9081fe9214af8f240d8459431972673
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2764142
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Auto-Submit: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>