Set 10 seconds of timeout for power.TurnOnDisplay.
By seeing the failures due to power.TurnOnDisplay, it looks like
it takes a very long time -- usually the entire timeout of the
test case -- for this failure.

Usually power.TurnOnDisplay should not take that long, and this
would increase the length of the entire test running time when
this happens. So this CL caps the limit for power.TurnOnDisplay,
and so this test fails in a more timely manner.

BUG=chromium:1188860
TEST=none

Change-Id: Ide98318267e84163779a6d931ed14e13e5a81d20