(Nearby Share) Add 'Some contacts' versions of CB->CB tests
Add subtests with 'Some contacts' visibility, where the receiver
explicitly chooses which contacts to be visible to for sharing. Added
new fixtures to support this, and added an argument for the sender's
username to the CrOSSetupRequest so we can configure the visibility
settings with the local service.

BUG=chromium:1185940
TEST=tast run -var=secondaryTarget=<dut2> <dut1> nearbyshare.*
TEST=tast run ... nearbyshare.CrosToCros*somecontacts*

Change-Id: Ic4d783ddf5b71de014e16f167547c3cf847beaae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2766166
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>