tast-tests: Disable Play Store for chrome.ARCEnabled
When a test relies on Play Store, Chrome should be booted with
chrome.ARCSupported(). Otherwise, chrome.ARCEnabled() is used. When
chrome.ARCEnabled() is used, Play Store does not function. Yet it can
still generate ramdom notifications (e.g. "Play Protect is enabled")
and they can interfere with UIAutomator and introduce flakiness.

As having Play Store for ARCEnabled() provides little meaningful
coverage, this CL disables the Play Store package when it's booted
with ARCEnabled. arc.IMEBlockingVK is doing that on its own but
it should be done in the Tast framework side.

BUG=b:154288462
TEST=tast run rammus-arc-r arc.IMEBlockingVK.vm
TEST=tast run rammus-arc-r arc.PlayStore.vm

Change-Id: I8bcd760a68bec99de2e58cd596f44647abc1eded