video.PlaybackPerf: Mute video.
Doing so ensures that the video doesn't stall because of issues in the
audio stack. To mute the video, we simply use video.html (which already
mutes its <video> element) instead of making Chrome visit the video file
directly. Note that we also set the <video>'s CSS max-height and
max-width to 100% so that the video is scaled to the size of the Chrome
window (which emulates the behavior before this CL).

BUG=b:182496681,b:182216163
TEST=video.PlaybackPerf.* on eldrid

Change-Id: I7b697b337929f9ed045323fdf2e163b2aefeb32b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2770281
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>