netlisten: Ignore all Tast-related processes
In crrev.com/c/2751503 we're going to randomize the ephemeral devserver
port. The CL fails on security.NetworkListeners test because its fixed
port number (28028) is hard-coded in the allowlist.

This CL updates the netlisten package to ignore all processes that is
started within the same SSH session as the current process.

BUG=chromium:1179179
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run $DUT security.NetworkListenersARC

Change-Id: Id9dceede5eb9c930cb35854c232dee85cff4c43a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2772381
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>