policy: Add QuickUnlockModeWhitelist tast test
Add the policy.QuickUnlockModeWhitelist tast test, which verifies that
the deprecated policy QuickUnlockModeWhitelist still behaves as
expected.

This policy was deprecated with M87 in favor of the more inclusive
QuickUnlockModeAllowlist, however we will continue support for the
deprecated policy and therefore want to verify that it still works
properly.

Once the support for this deprecated policy is stopped, this test can be
removed.

BUG=chromium:1187464
TEST=tast run <DUT> policy.QuickUnlockModeWhitelist

Change-Id: I4e4325faa5254abec3543569f0c642ccf55bfb37