Tast: Run CheckIntelSARTable on Akemi
Currently, Akemi is being skipped by the critical version of
wifi.CheckIntelSARTable. The issue causing the test to fail on Akemi has
been resolved, and Akemi is currently passing the informational version
of the test, so make the critical version of the test run on Akemi
again.

BUG=b:172288894
TEST=tast run ${DUT} wifi.CheckIntelSARTable
passes on Akemi
test run ${DUT} wifi.CheckIntelSARTable.informational
skips on Akemi

Change-Id: I2dac0ea4be38aa6e8d4bb570f37211aecc2d7dd6