tast-tests: typec: Update ModeReboot to use smoke test pattern
Modify ModeReboot to use the same pattern for its CQ runs as
ModeCrash, which is to add a boolean attribute to the param structs
which signifies it is a smoke test.

To do that, modify the checkPortForTBTPartner to check for all ports
instead of just one. While we are here, parse INVALID_PARAM similar to
the other tests, and return false instead of returning an error.

BUG=b:179338742
TEST=-Try typec.ModeReboot.smoke locally; verify that it skips early
     (whether a TBT dock is connected or not).
     -Try typec.ModeReboot.manual locally; verify that it behaves
     as expected (whether a TBT dock is connected or not).

Change-Id: I8afbf145ec2b949a7add384f74e0647297bedb7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2773481
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>