tast-tests: Add method to maximize Files app if action fails
Files app has a race condition that causes the listbox to not populate
on rare occasions. Add a method which attempts the supplied action and
if it fails, maximizes the Files app window and tries the action again.

The act of maximizing the window causes the event handlers to redraw the
listbox and repopulate it. This has been a clear blocker for getting
tests on the CQ and so having a way to stabilize these tests (until a
fix is identified) will achieve that.

BUG=chromium:1189914
TEST=./fast_build.sh -T and follow up CL

Change-Id: Icfc978d3cf314ccddd94b74df4c8ce35069aa670