tast-tests: Maximize window if file is not found
One of the last known flakes to occur in the filemanager.DrivefsUI test
has been an empty listbox. This was identified as a race condition and
is not the core purpose of the test, maximizing the window causes the
event listeners to kick off again and repopulate the listbox.

Update the filemanager.DragDrop test to the new uiauto library and
perform a similar change as the test is exhibiting the same flaky
behaviour. Have left a TODO to remove this workaround once the
underlying race condition has been fixed.

BUG=chromium:1177053, chromium:1189914, chromium:1173532
TEST=tast -verbose run <DUT-IP> filemanager.DrivefsUI
TEST=tast -verbose run <DUT-IP> filemanager.DragDrop

Change-Id: I71edf89c647eec105ecd212028a9d9d5d5416d75
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2772695
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: Austin Tankiang <austinct@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>