tast-tests: Enable ChromeVox hint for dev mode.
This is a follow-up to the following change:
https://chromium-review.googlesource.com/c/chromium/src/+/2747223

The above change disables the ChromeVox hint by default when the
device is in dev mode. To ensure we maintain test coverage, enable
the ChromeVox hint in dev mode for tast tests by adding the necessary
command line flag.

BUG=b:182245799
TEST=N/A
AX-Relnotes: N/A

Change-Id: Ie4c667ba6d98302cf25ceb974e2c632bfe604a3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2746794
Tested-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>