tast-tests: Resize DragDropActivity from Tast
Currently, arc.DragDrop uses StartupActivity to show an activity in a
certain location. It can be made simpler by updating the window bounds
from golang Tast code.

Also, minimum window size in ARC R is larger than 300x300 This CL
changes the sizes of the activity and the extension window to 500x500.

BUG=b:183394117
TEST=tast run rammus-arc-r-DUT arc.DragDrop.vm

Change-Id: I6c3d9d5eec436a153a11307cb8665c1428687435