Fix ui.LauncherDragPerf flakiness
When running with ui.HotseatAnimation.overflow_shelf, I noticed
that the shelf icon bounds calculation in ui.LauncherDragPerf
can be unstable. When an old bounds are used, it may end up with
dragging incorrect bounds and fail.

To prevent this case, this CL adds a waiting so that the shelf
icon becomes stable.

BUG=none
TEST=tast run ui.HotseatAnimation.overflow_shelf ui.LauncherDragPerf

Change-Id: Iccf47191c91dcf309241ec59014ba73da861192b