wifi: Use subtest instead of params for ChannelScanDwellTime
Using params results in the test being conceptually three
distinct tests with separate entries in Stainless and in
crosbolt. This should instead be a single test in Stainless
and the performance numbers should be separate graphs on the
same page.

See also crrev.com/c/2746354 which parameterized the test to
run on multiple channels.

I've verified that the test is still working with this change,
the test name reverts from `wifi.ChannelScanDwellTime.ch<N>`
back to a single `wifi.ChannelScanDwellTime`, and produces a
single `results-chart.json` containing three metrics (rather
than the three separate JSON files with one metric each that
Params gets us).

So this ought to behave as desired on Stainless/crosbolt.

BUG=b:181294019
TEST=`tast run $DUT wifi.ChannelScanDwellTime` on some DUTs

Change-Id: Ia5a3b3af732a46ea09f259934f0ad4ef44f361de