tast-tests: allow option to configure router as pcap
This was originally only available as a fallback when a pcap host wasn't
available in the test setup. Allow it to be configured as an option so
that we can get a 2nd party capture.

BUG=b:180664130,b:183240266
TEST=wrote a custom test, used the option, looked at pcap and saw Assoc
Request frames that aren't available in 3rd party capture.

Change-Id: I4ca75a97cadf5c1cf12c09c3b69ac1199ccde98c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2783567
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@chromium.org>