tast-tests: Disable apps.SystemWebAppsInstallGuest*
The test apps.SystemWebAppsInstallGuest* uses a guest session to verify
that the SWAs are all correctly installed. It seems that the Tast
framework makes an attempt at reusing the Chrome session when it can and
for a subsequent test (crash.Sender.mock_consent) this ends up reusing
the guest mode which causes it to fail.

Disabling apps.SystemWebAppsInstallGuest* until we either have the
individual test clean up the guest session, the fixture clean it up or
have each test ensure it has the correct Chrome instance running.

BUG=chromium:1190865
TEST=tast -verbose <amd64-generic-DUT> apps.SystemWebAppsInstallGuest*
crash.Sender.mock_consent

Change-Id: Ib37dbaedd9dfdfad4041142093292777edc74575